Add polyfills and integrate across project
