+++ index.html
@@ -91,6 +91,7 @@
   >
     <div class="section-skeleton" aria-hidden="true"></div>
   </section>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
 
 
   <script src="/content/webentwicklung/utils/debug-config.js" defer></script>
