+++ pages/album.html
@@ -13,6 +13,7 @@
     <h1>Album</h1>
     <p>Album-Seite – Inhalte folgen.</p>
   </main>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
 </body>
