+++ pages/features/wetter.html
@@ -13,6 +13,7 @@
     <h1>Wetter</h1>
     <p>Feature-Seite – Inhalte folgen.</p>
   </main>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
 </body>
