+++ pages/fotogalerie/urban.html
@@ -13,6 +13,7 @@
     <h1>Urban Exploration</h1>
     <p>Galerieseite – Inhalte folgen.</p>
   </main>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
 </body>
