+++ pages/spiele/space-defender.html
@@ -13,6 +13,7 @@
     <h1>Space Defender</h1>
     <p>Spieleseite – Inhalte folgen.</p>
   </main>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
 </body>
