+++ pages/ueber-mich/index.html
@@ -43,6 +43,7 @@
   </main>
 
   <div id="footer-placeholder"></div>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
 
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
