+++ pages/webentwicklung/project-1.html
@@ -13,6 +13,7 @@
     <h1>E-Commerce Platform</h1>
     <p>Projektseite – Details folgen.</p>
   </main>
+  <script src="/content/webentwicklung/utils/polyfills.js" defer></script>
   <script src="/content/webentwicklung/main.js" defer></script>
   <script src="/content/webentwicklung/menu/menu.js" defer></script>
 </body>
